# packages/shared/openapi.yaml
openapi: 3.1.0
info:
  title: AI Inference Gateway
  version: 0.6.0
  description: OpenAI-shaped chat completions edge for multi-cloud LLM routing.
servers:
  - url: http://localhost:18080
paths:
...
  /v1/chat/completions:
    post:
      summary: Create a chat completion
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/CacheBypass"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/chat-completion.request.json"
      responses:
        "200":
          description: |
            JSON completion when `stream` is false/omitted.
            When `stream` is true, OpenAI-shaped SSE (`text/event-stream`) with
            `chat.completion.chunk` events ending in `data: [DONE]`.
            Soft budget warnings may include header `X-Budget-Warning: soft`.
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        "400":
          description: Invalid request body
        "401":
          description: Missing or invalid API key
        "402":
          description: Tenant hard USD/token budget exhausted (configurable as 429)
        "413":
          description: Request body exceeds MAX_BODY_BYTES
        "429":
          description: Tenant hard budget exhausted when BUDGET_HARD_STATUS=429
        "502":
          description: Upstream router / provider error
components:
  parameters:
...
    CacheBypass:
      name: X-Cache-Bypass
      in: header
      required: false
      description: Set to "1" to skip semantic cache read/write
      schema:
        type: string
        enum: ["1", "true"]
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Maps to a tenant via TENANT_API_KEYS (or DEMO_API_KEY → tenant "default").